client/requirements.txt
elevenlabs
streamlit==1.36.0
python-dotenv
google-cloud-firestore